fix warnings of possible overflow after multiplication

+++ src/render/opengl/SDL_render_gl.c
@@ -1478,7 +1478,7 @@ static int GL_RenderReadPixels(SDL_Renderer *renderer, const SDL_Rect *rect,
     }
 
     temp_pitch = rect->w * SDL_PIXELFORMAT_BPP(temp_format);
-    temp_pixels = SDL_malloc((rect->h + (renderer->target ? 0 : 1)) * temp_pitch);
+    temp_pixels = SDL_malloc((size_t)(rect->h + (renderer->target ? 0 : 1)) * temp_pitch);
     if (!temp_pixels) {
         return SDL_OutOfMemory();
     }
